+++ Changelog.md
@@ -23,6 +23,7 @@
 - Place comments from the middle of the call to the end
 - Drop line breaks between operator arguments (except '+', "-", "&&" and "||")
 - Add non-null assertions on call site for non-null parameters
+- Specify type for variables with anonymous type if they have write accesses
 
 ### IDE
 - Debugger can distinguish nested inline arguments